I can draw boxes now

+++ src/window/opengl/shaders/pixels.vert.glsl
@@ -1,7 +1,7 @@
 #version 330 core
 
-layout(location = 0) in vec2 a_Position;
-layout(location = 1) in vec2 a_TexCoords;
+layout(location = 0) in uvec2 a_Position;
+layout(location = 1) in uvec2 a_TexCoords;
 layout(location = 2) in vec4 a_Color;
 layout(location = 3) in uint a_TextureID;
